[dart2js] Modify error parsing to support new d8 error/stacktrace formatting.
d8 used to include the error message between the stack trace and the error location indicator ("^").

With the newest version being updated to in https://dart-review.googlesource.com/c/sdk/+/362386 the error is no longer located there. It is instead found toward the beginning of the error message.

Both the new and old version of d8 have the error in that location (the old version just duplicated it into the spot the test currently reads it from).

Change-Id: I15143510cd9ed6b8434a5531a08011854ce2da9c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/362580
Auto-Submit: Nate Biggs <natebiggs@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>