Move all calls to FlowAnalysis.getMatchedValueType to TypeAnalyzer.
This avoids redundant computation of the matched value type, so it
should lead to a modest (but perhaps negligible) performance
improvement.

It also paves the way for some future work I'm planning that will
allow details of the type inference process to be logged for debugging
purposes.

Change-Id: I8fb24a5b07b97f17ed045bc8bd2ca9e854233b20
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/364800
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>