Add explicit indicators for diagnostics that are missing documentation
I believe that this will make it easier to identify which diagnostics
are undocumented (though I haven't done this for parser-produced
diagnostics). It will also allow us to easily count how many there are
in order to monitor progress / regress.

I also normalized the location of the flag to make the file a little
cleaner.

Change-Id: I4fe07f446508788a8c02efcd8e78e94689ffb40b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/365283
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>