[dart2wasm] Make use of TFA-inferred types in more cases
Avoid falling back to variable type if the variable is `final` and has a
more precise inferred type.

Use unboxing information on methods to make unboxed return types
when TFA concludes it's int/double.

This will partially fix a performance regression in SplayHarder
benchmark where a parameter of type `num` was unboxed to `double` but
the local variable was kept as `num` and therefore always caused a box
allocation. After this change the box allocation is moved to the place
where we call `double.toString()`.

TEST=Updated expectation files.

Change-Id: I0252cf86ada9a8affbb46a31e913504930057650
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/365560
Reviewed-by: Slava Egorov <vegorov@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>