[linter] Amend invalid_runtime_check_with_js_interop_types to catch additional is checks
In the case where the is check is trivially true because the
erased type of the left type is a subtype of the erased type
of the right type, it may be confusing for users who are used
to dart:html-style checks, which checked the JS type of the
value. In this case, we should warn users that this is
trivially true and that they might not be doing the is check
they think they are doing.

Change-Id: I156915a5e13acb67a4096b7cab4426d8aa4ed15f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/364167
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Srujan Gaddam <srujzs@google.com>