Move LinterContext.resolveNameInScope to a utility function
This utility function is moved to the linter package, closer to the
two callers (two lint rules). The utility function takes with it,
the `LinterNameInScopeResolutionResult` class and
`_LinterNameInScopeResolutionResultState` enum. The implementation is
great, so it's not changed, and the existing API is good too.

Change-Id: I7809661e67ae38fd149466eea6bb45b1b632d7d9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/365663
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>