Set type of TypeLiteral in ResolverVisitor for consistency.
TypeLiteral AST nodes are not created during parsing; they are created
either during identifier resolution (ResolutionVisitor) or during full
resolution (ResolverVisitor) as part of an AST rewriting process.

Before this change, the static type of TypeLiteral nodes was always
set at the time of creation; this was possible because TypeLiteral
nodes always have a static type of `Type`. But it was inconsistent,
because in most cases, the ResolverVisitor is responsible for
assigning static types to AST nodes.

This change shifts the responsibility for setting the static types of
TypeLiteral nodes to the ResolverVisitor, eliminating the
inconsistency.

This helps pave the way for some code I'm working on that will allow
the analyzer's resolution process to be instrumented. That code will
perform assertions to make sure that the ResolverVisitor assigns a
static type to every expression it visits (this is needed to make sure
that the instrumentation logic is correct), so eliminating the
inconsistency now will avoid a future assertion failure.

Change-Id: I740cee6e196638af73dc20b2de22015ea2dc4c0d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/365921
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>