[cfe] Remove LibraryBuilder.isNonNullableByDefault
This removes the `LibraryBuilder.isNonNullableByDefault` property, which
is now always `true`, and with it all expressions that directly used its
value. Calls which passes this value use `true` instead. It is further
work to remove the parameters that have now become trivial and also
the error messages that are no longer used.

Change-Id: Ib7441263022cdd9f251dfcec32d51e26f2732161
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/365807
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>