[wildcards] update `UNUSED_LOCAL_VARIABLE` for wildcards
Updates `UNUSED_LOCAL_VARIABLE` to flag `__`s when the wildcard variables feature is enabled.

Fixes: https://github.com/dart-lang/sdk/issues/55719

The implementation is quick so I'm very open to feedback.

Also, I'm not sure how exhaustively we want/need to test. Coverage is incomplete and I'm not sure it needs to be *but* for sure if there are interesting cases missed, please do let me know!



Change-Id: I4d95d5b94141fa033b3c426ac4b5a721a12fc107
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/366403
Reviewed-by: Kallen Tu <kallentu@google.com>
Commit-Queue: Phil Quitslund <pquitslund@google.com>