Remove unnecessary Linter parent class, rename linterRegistry
Linter was in between LintRule and NodeLintRule, but the only subtype
was LintRule so the abstraction was unnecessary. There are lots of
small, nice follow-on effects across the codebase.

Also rename the top-level `lintRegistry` variable to be
`lintRuleTimers` as it is just a mapping of lint rule timers. (I was
confused about what this was because `Analyzer.registeredRules` is the
more general "registry of rules."

Also remove the deprecated, unused `getVisitor` API.

Also remove the unused top-level `linterVersion` string.

Cq-Include-Trybots: luci.dart.try:flutter-analyze-try,analyzer-win-release-try,pkg-win-release-try
Change-Id: Icf75db7786fc2022e271ec6f32fd3aa00b2b4084
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/366840
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>