[dart2wasm] Use the `main` method from `Component` instead of (arbitrarily) assuming the first library is the main library
The CFE compilation should fail if there is no `main`, so dart2wasm
compiler should be able to assume it's there in
`Component.mainMethod`.

This fixes an issue when the `main` method in the main library isn't
there but rather re-exported from another library (which is not the
first one).

(This is part of work towards adding a CI builder using
`dart compile wasm`)

Change-Id: If9e75c2bf1c175ec9db1d97bd56b00cfdb6d3cb9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/366800
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>