Remove the not-imported contributor
The local library contributor can't be removed yet because it is still
being used by the Cider support, but other than that this is the last of
the contributors.

It isn't clear to me whether the local library contributor is still of
value for Cider or whether it was correcting for the local suggestions
that were, for unknown reasons, being contributed by the not-imported
contributor. That will require some investigation.

Change-Id: I7d3235ed19af67bee87f1dc9af38488bfdd6169b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/366968
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>